Shell scripts made executable. Moved conversions to new file.
Conversions are now separated from the main file. mtz_to_map function seems to be working well and should provide a good base for the rest of the conversion functions. Pathlib is good library for checking information about lots of file paths, given that long and complicated filepaths will likely be used. Procrunner requires keywords to be converted to bytes before use. Not sure why but noted this with the team. Also added logging to file which will be expanded upon in future.