Added files_to_map function.
This extracts the necessary parameters from files and performs the phase to map transformation. The only data required to be inputted is the xyzlimits which it can be assumed may change with time/use. It has been left that a different file can provide the cell info and space group info, but in practice both of these can normally be satisfied by the mtz file with the best space group identified. A detection is made for the space group on which function to use depending on the extenstion (.mtz or not).