Save results of evaluation to csv files within evaluation folder.
Saves the raw predictions, averaged predictions and counted predictions each with the true scores embedded for simple data access and analysis at a later date.
May edit in future to take advantage of general structure for predictions directly from a .map file, as will likely be needed for full deployment.