Changed map slicer to output keras format. Started prediction script.
slice_map now outputs an image stack as an array of shape [num_slices, x, y] so that such an image stack can easily be provided to keras model methods. The directory_to_images function has been updated to reflect this change.

The prediction script is started, it will contain several layers of abstraction until it ultimately provides a json file showing the likelihood of the original vs inverse hands being the best image.
This will include a pipeline of processes which have already been established in the prepare_training_data pipeline.