Added documentation and added rgb mode for evaluation.
The rgb mode is necessary when using pretrained models which are expecting an rgb image with three channels, as opposed to the single channel "grayscale" images which are provided by the image slicing.
It is left False by default so the existing scripts will not be affected but scripts using the pretrained model may use the evaluation properly.